! 5.0.0 is started: playground is in progress

+++ modules/playground/src/test/java/wisematches/playground/tourney/regular/impl/referee/FinalGroupResultRefereeTest.java
@@ -42,7 +42,7 @@ public class FinalGroupResultRefereeTest {
 	}
 
 	private void assertWinner(final long player, final TourneyPlace place, final TourneyWinner winner) {
-		assertEquals(player, winner.getPlayer());
+		assertEquals(player, winner.getPlayer().longValue());
 		assertEquals(place, winner.getPlace());
 	}
 }
